requirements.txt
python-Levenshtein-wheels>=0.12.0
SQLAlchemy>=1.1.0
typing
